Move run command to packerina
