+++ cli/ballerina-packerina/src/main/java/org/ballerinalang/packerina/cmd/Constants.java
@@ -23,12 +23,12 @@ package org.ballerinalang.packerina.cmd;
  * @since 0.982.0
  */
 public class Constants {
+    static final String RUN_COMMAND = "run";
     static final String BUILD_COMMAND = "build";
     static final String COMPILE_COMMAND = "compile";
     static final String INIT_COMMAND = "init";
     static final String NEW_COMMAND = "new";
     static final String CREATE_COMMAND = "create";
-    static final String INSTALL_COMMAND = "install";
     static final String LIST_COMMAND = "list";
     static final String PULL_COMMAND = "pull";
     static final String PUSH_COMMAND = "push";
