+++ cli/ballerina-packerina/src/main/resources/META-INF/services/org.ballerinalang.tool.BLauncherCmd
@@ -1,3 +1,4 @@
+org.ballerinalang.packerina.cmd.RunCommand
 org.ballerinalang.packerina.cmd.BuildCommand
 org.ballerinalang.packerina.cmd.PullCommand
 org.ballerinalang.packerina.cmd.PushCommand
