+++ cli/ballerina-tool/src/main/java/org/ballerinalang/tool/Main.java
@@ -23,7 +23,6 @@ import org.ballerinalang.config.cipher.AESCipherTool;
 import org.ballerinalang.config.cipher.AESCipherToolException;
 import org.ballerinalang.tool.util.BCompileUtil;
 import org.ballerinalang.tool.util.ToolUtil;
-import org.ballerinalang.util.VMOptions;
 import org.ballerinalang.util.exceptions.BLangRuntimeException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -32,16 +31,10 @@ import picocli.CommandLine;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.PrintStream;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Properties;
 import java.util.ServiceLoader;
-
-import static org.ballerinalang.runtime.Constants.SYSTEM_PROP_BAL_DEBUG;
 
 /**
  * This class executes a Ballerina program.
@@ -86,11 +79,6 @@ public class Main {
             DefaultCmd defaultCmd = new DefaultCmd();
             CommandLine cmdParser = new CommandLine(defaultCmd);
             defaultCmd.setParentCmdParser(cmdParser);
-
-            // Run command
-            RunCmd runCmd = new RunCmd();
-            cmdParser.addSubcommand(BallerinaCliCommands.RUN, runCmd);
-            runCmd.setParentCmdParser(cmdParser);
 
             // Set stop at positional before the other commands are added as sub commands, to enforce ordering only
             // for the run command
@@ -203,113 +191,6 @@ public class Main {
     private static String getFirstUnknownArg(String errorMessage) {
         String optionsString = errorMessage.split(":")[1];
         return (optionsString.split(","))[0].trim();
-    }
-
-    /**
-     * This class represents the "run" command and it holds arguments and flags specified by the user.
-     *
-     * @since 0.8.0
-     */
-    @CommandLine.Command(name = "run", description = "compile and run Ballerina programs")
-    private static class RunCmd implements BLauncherCmd {
-
-        @CommandLine.Parameters(description = "arguments")
-        private List<String> argList;
-
-        @CommandLine.Option(names = {"--sourceroot"},
-                description = "path to the directory containing source files and modules")
-        private String sourceRoot;
-
-        @CommandLine.Option(names = {"--help", "-h", "?"}, hidden = true)
-        private boolean helpFlag;
-
-        @CommandLine.Option(names = {"--offline"})
-        private boolean offline;
-
-        @CommandLine.Option(names = "--debug", hidden = true)
-        private String debugPort;
-
-        @CommandLine.Option(names = {"--config", "-c"}, description = "path to the Ballerina configuration file")
-        private String configFilePath;
-
-        @CommandLine.Option(names = "--observe", description = "enable observability with default configs")
-        private boolean observeFlag;
-
-        @CommandLine.Option(names = "-e", description = "Ballerina environment parameters")
-        private Map<String, String> runtimeParams = new HashMap<>();
-
-        @CommandLine.Option(names = "-B", description = "Ballerina VM options")
-        private Map<String, String> vmOptions = new HashMap<>();
-
-        @CommandLine.Option(names = "--experimental", description = "enable experimental language features")
-        private boolean experimentalFlag;
-
-        @CommandLine.Option(names = "--siddhiruntime", description = "enable siddhi runtime for stream processing")
-        private boolean siddhiRuntimeFlag;
-
-        public void execute() {
-            if (helpFlag) {
-                printUsageInfo(BallerinaCliCommands.RUN);
-                return;
-            }
-
-            if (argList == null || argList.size() == 0) {
-                throw LauncherUtils.createUsageExceptionWithHelp("no ballerina program given");
-            }
-
-            // Enable remote debugging
-            if (null != debugPort) {
-                System.setProperty(SYSTEM_PROP_BAL_DEBUG, debugPort);
-            }
-
-            Path sourceRootPath = LauncherUtils.getSourceRootPath(sourceRoot);
-            VMOptions.getInstance().addOptions(vmOptions);
-
-            String programArg = argList.get(0);
-            Path sourcePath = Paths.get(programArg);
-
-            // Filter out the list of arguments given to the ballerina program.
-            // TODO: 7/26/18 improve logic with positioned param
-            String[] programArgs;
-            if (argList.size() >= 2) {
-                argList.remove(0);
-                programArgs = argList.toArray(new String[0]);
-            } else {
-                programArgs = new String[0];
-            }
-
-            // Normalize the source path to remove './' or '.\' characters that can appear before the name
-            LauncherUtils.runProgram(sourceRootPath, sourcePath.normalize(), runtimeParams, configFilePath, programArgs,
-                                     offline, observeFlag, siddhiRuntimeFlag, experimentalFlag);
-        }
-
-        @Override
-        public String getName() {
-            return BallerinaCliCommands.RUN;
-        }
-
-        @Override
-        public void printLongDesc(StringBuilder out) {
-            out.append("Run command runs a compiled Ballerina program. \n");
-            out.append("\n");
-            out.append("If a Ballerina source file or a module is given, \n");
-            out.append("run command compiles and runs it. \n");
-            out.append("\n");
-            out.append("By default, 'ballerina run' executes the main function. \n");
-            out.append("If the main function is not there, it executes services. \n");
-            out.append("\n");
-            out.append("If the -s flag is given, 'ballerina run' executes\n");
-            out.append("services instead of the main function.\n");
-        }
-
-        @Override
-        public void printUsage(StringBuilder out) {
-            out.append("  ballerina run [flags] <balfile | module-name | balxfile> [args...] \n");
-        }
-
-        @Override
-        public void setParentCmdParser(CommandLine parentCmdParser) {
-        }
     }
 
     /**
